requirements.txt
pandas==1.5.3
tornado==6.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
Jinja2==3.1.2
Flask==2.2.2
python-dotenv==0.21.1